[#36] Farcaster mini app manifest and official wagmi connector
- Add public/.well-known/farcaster.json manifest (accountAssociation
  placeholder — operator must fill with domain signature)
- Install @farcaster/miniapp-wagmi-connector, replace custom connector
- Extract isFarcasterMiniApp() into lib/farcaster-detect.ts
- Update ConnectWallet to use official connector type "farcasterMiniApp"
- Remove custom lib/farcaster-connector.ts (replaced by official package)

Existing FarcasterMiniApp component (sdk.actions.ready()) unchanged.

Fixes #36
