+++ public/.well-known/farcaster.json
@@ -1,0 +1,18 @@
+{
+  "accountAssociation": {
+    "header": "",
+    "payload": "",
+    "signature": ""
+  },
+  "miniapp": {
+    "version": "1",
+    "name": "PlotLink",
+    "iconUrl": "https://plotlink.xyz/favicon.ico",
+    "homeUrl": "https://plotlink.xyz",
+    "splashBackgroundColor": "#0a0a0a",
+    "description": "Tokenise your story from day 1. Publish plots, drive trading, earn royalties from every trade.",
+    "primaryCategory": "entertainment",
+    "tags": ["stories", "writing", "onchain", "base"],
+    "requiredChains": ["eip155:8453"]
+  }
+}
